+++ forest/benchmarking/operator_tools/project_superoperators.py
@@ -11,11 +11,12 @@ A good reference for these methods is:
       https://arxiv.org/abs/1803.10062
 """
 import numpy as np
+from scipy import linalg
 from forest.benchmarking.utils import partial_trace
-from forest.benchmarking.operator_tools.superoperator_transformations import vec
+from forest.benchmarking.operator_tools.superoperator_transformations import vec, unvec, kraus2choi
 
 
-def proj_choi_to_completely_positive(choi: np.ndarray) -> np.ndarray:
+def proj_choi_to_completely_positive(choi: np.ndarray, check_finite: bool = True) -> np.ndarray:
     """
     Projects the Choi representation of a process into the nearest Choi matrix in the space of
     completely positive maps.
@@ -29,10 +30,11 @@ def proj_choi_to_completely_positive(choi: np.ndarray) -> np.ndarray:
           https://arxiv.org/abs/1803.10062
 
     :param choi: Choi representation of a process
+    :param check_finite: check that the input matrices contain only finite numbers.
     :return: closest Choi matrix in the space of completely positive maps
     """
-    hermitian = (choi + choi.conj().T) / 2  # enforce Hermiticity
-    evals, v = np.linalg.eigh(hermitian)
+    hermitian_choi = (choi + choi.conj().T) / 2  # enforce Hermiticity
+    evals, v = linalg.eigh(hermitian_choi, check_finite=check_finite)
     evals[evals < 0] = 0  # enforce completely positive by removing negative eigenvalues
     diag = np.diag(evals)
     return v @ diag @ v.conj().T
@@ -53,7 +55,7 @@ def proj_choi_to_trace_non_increasing(choi: np.ndarray) -> np.ndarray:
     pt = partial_trace(choi, dims=[dim, dim], keep=[0])
 
     hermitian = (pt + pt.conj().T) / 2  # enforce Hermiticity
-    d, v = np.linalg.eigh(hermitian)
+    d, v = linalg.eigh(hermitian)
     d[d > 1] = 1  # enforce trace preserving
     D = np.diag(d)
     projection = v @ D @ v.conj().T
@@ -143,3 +145,35 @@ def proj_choi_to_physical(choi: np.ndarray, make_trace_preserving: bool = True) 
         last_state = new_state
 
     return new_state
+
+
+def proj_choi_to_unitary(choi: np.ndarray, check_finite: bool = True) -> np.ndarray:
+    """
+    Compute the unitary closest to a quantum process specified by a Choi matrix.
+
+    This function enforces Hermiticity of the Choi matrix. See the following reference for more
+    details:
+
+    Interference of Quantum Channels
+    Daniel K. L. Oi
+    Phys. Rev. Lett. 91, 067902 (2003)
+    https://doi.org/10.1103/PhysRevLett.91.067902
+    https://arxiv.org/abs/quant-ph/0303178
+
+    :param choi: the Choi representation of a quantum process.
+    :param check_finite: check that the input matrices contain only finite numbers.
+    :return: choi matrix corresponding to the closest unitary
+    """
+    dim = int(np.sqrt(choi.shape[0]))
+    hermitian_choi = (choi + choi.conj().T) / 2  # enforce Hermiticity
+    vals, vs = linalg.eigh(hermitian_choi, check_finite=check_finite)
+
+    # vec corresponding to largest eval =  Kraus OP with largest norm
+    large_eigen_vec = vs[:, np.argmax(vals)].reshape((dim * dim, 1))
+    kraus = unvec(large_eigen_vec)
+
+    U, s, V = linalg.svd(kraus)
+    unitary = U @ V
+    # pick a global phase convention, we choose the first element
+    phase = np.angle(unitary[0, 0])
+    return kraus2choi(np.exp(-1j * phase) * unitary)